# docs/学生在线纳新系统 API.openapi.yaml
openapi: 3.1.0
info:
  title: 学生在线纳新系统 API
  description: 学生在线纳新系统后端接口
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/departments/{departmentId}/applications/{applicationId}/admission:
    put:
      tags:
      - 部门报名
      operationId: admit
      parameters:
      - name: departmentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      - name: applicationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultDepartmentApplicationVO"
      security:
      - sessionCookie: []
    delete:
      tags:
      - 部门报名
      operationId: cancelAdmissionDraft
      parameters:
      - name: departmentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      - name: applicationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultDepartmentApplicationVO"
      security:
      - sessionCookie: []
...
  /api/departments/{departmentId}/applications:
    get:
      tags:
      - 部门报名
      summary: 分页查询部门报名信息
      operationId: findApplications
      parameters:
      - name: campus
        in: query
        required: false
        description: 学生所在校区；不传则不筛选，包含未填写校区的历史用户。
        schema:
          type: string
          enum: [SOFTWARE_PARK, CENTRAL, QIANFOSHAN, XINGLONGSHAN, HONGJIALOU, BAOTUQUAN]
      - name: departmentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      - name: keyword
        in: query
        required: false
        schema:
          type: string
          maxLength: 100
          minLength: 0
      - name: college
        in: query
        required: false
        schema:
          type: string
          maxLength: 64
          minLength: 0
      - name: grade
        in: query
        required: false
        schema:
          type: integer
          format: int32
          maximum: 2100
          minimum: 2000
      - name: interviewed
        in: query
        required: false
        schema:
          type: boolean
      - name: status
        in: query
        required: false
        description: 报名状态；不传则查询全部状态
        schema:
          type: string
          enum:
          - SUBMITTED
          - INTERVIEW_COMPLETED
          - ADMISSION_DRAFT
          - ADMITTED
      - name: sortBy
        in: query
        required: false
        schema:
          type: string
          default: submittedAt
          pattern: submittedAt|score
      - name: sortOrder
        in: query
        required: false
        schema:
          type: string
          default: desc
          pattern: asc|desc
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
          minimum: 1
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultPageVODepartmentApplicationSummaryVO"
      security:
      - sessionCookie: []
    post:
      tags:
      - 部门报名
      summary: 提交部门报名
      operationId: submit
      parameters:
      - name: departmentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentApplicationRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultMyDepartmentApplicationVO"
      security:
      - sessionCookie: []
  /api/departments/{departmentId}/applications/admissions/publish:
    post:
      tags:
      - 部门报名
      operationId: publishAdmissions
      parameters:
      - name: departmentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdmissionPublishRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultAdmissionPublishVO"
      security:
      - sessionCookie: []
...
  /api/departments/{departmentId}/applications/{applicationId}:
    get:
      tags:
      - 部门报名
      summary: 获取报名记录详情
      operationId: getApplicationDetail
      parameters:
      - name: departmentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      - name: applicationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultDepartmentApplicationDetailVO"
      security:
      - sessionCookie: []
  /api/departments/{departmentId}/applications/me:
    get:
      tags:
      - 部门报名
      operationId: findMyApplication
      parameters:
      - name: departmentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultMyDepartmentApplicationVO"
      security:
      - sessionCookie: []
    delete:
      tags:
      - 部门报名
      summary: 取消当前用户的报名
      description: 仅允许取消 SUBMITTED 且尚未签到、进入面试或录取流程的报名。
      operationId: cancelMyApplication
      parameters:
      - name: departmentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultVoid"
      security:
      - sessionCookie: []
  /api/departments/{departmentId}/applications/export:
    get:
      tags:
      - 部门报名
      summary: 导出部门报名信息
      operationId: exportApplications
      parameters:
      - name: campus
        in: query
        required: false
        description: 学生所在校区；不传则不筛选，包含未填写校区的历史用户。
        schema:
          type: string
          enum: [SOFTWARE_PARK, CENTRAL, QIANFOSHAN, XINGLONGSHAN, HONGJIALOU, BAOTUQUAN]
      - name: departmentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          exclusiveMinimum: 0
      - name: keyword
        in: query
        required: false
        schema:
          type: string
          maxLength: 100
          minLength: 0
      - name: college
        in: query
        required: false
        schema:
          type: string
          maxLength: 64
          minLength: 0
      - name: grade
        in: query
        required: false
        schema:
          type: integer
          format: int32
          maximum: 2100
          minimum: 2000
      - name: interviewed
        in: query
        required: false
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
      security:
      - sessionCookie: []
...
components:
  schemas:
...
    DepartmentApplicationVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        departmentId:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - SUBMITTED
          - INTERVIEW_COMPLETED
          - ADMISSION_DRAFT
          - ADMITTED
        submittedAt:
          type: string
          format: date-time
...
    MyDepartmentApplicationVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        departmentId:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - SUBMITTED
          - INTERVIEW_COMPLETED
          - ADMITTED
        submittedAt:
          type: string
          format: date-time
    ResultMyDepartmentApplicationVO:
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          $ref: "#/components/schemas/MyDepartmentApplicationVO"
        msg:
          type: string
        timestamp:
          type: integer
          format: int64
    ResultDepartmentApplicationVO:
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          $ref: "#/components/schemas/DepartmentApplicationVO"
        msg:
          type: string
        timestamp:
          type: integer
          format: int64
...
    ApplicationAnswerRequest:
      type: object
      properties:
        questionId:
          type: integer
          format: int64
        answerText:
          type: string
          maxLength: 5000
          minLength: 0
        optionIds:
          type: array
          items:
            type: integer
            format: int64
          maxItems: 50
          minItems: 0
    DepartmentApplicationRequest:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/ApplicationAnswerRequest"
          maxItems: 100
          minItems: 0
    AdmissionPublishRequest:
      type: object
      properties:
        subject:
          type: string
          maxLength: 200
          minLength: 0
        content:
          type: string
          maxLength: 10000
          minLength: 0
      required:
      - content
      - subject
    AdmissionPublishVO:
      type: object
      properties:
        publishedCount:
          type: integer
          format: int32
    ResultAdmissionPublishVO:
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          $ref: "#/components/schemas/AdmissionPublishVO"
        msg:
          type: string
        timestamp:
          type: integer
          format: int64
...
    DepartmentApplicationSummaryVO:
      type: object
      properties:
        campus:
          description: 学生所在校区；历史资料可为 null。更新资料时缺省或 null 表示不修改。
          type: [string, "null"]
          enum: [SOFTWARE_PARK, CENTRAL, QIANFOSHAN, XINGLONGSHAN, HONGJIALOU, BAOTUQUAN, null]
        id:
          type: integer
          format: int64
        casId:
          type: string
        applicantName:
          type: string
        college:
          type: string
        major:
          type: string
        grade:
          type: integer
          format: int32
        phone:
          type: string
        qq:
          type: string
        status:
          type: string
          enum:
          - SUBMITTED
          - INTERVIEW_COMPLETED
          - ADMISSION_DRAFT
          - ADMITTED
        submittedAt:
          type: string
          format: date-time
        interviewed:
          type: boolean
        interviewId:
          type: integer
          format: int64
        score:
          type: number
          format: double
    PageVODepartmentApplicationSummaryVO:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DepartmentApplicationSummaryVO"
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
    ResultPageVODepartmentApplicationSummaryVO:
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          $ref: "#/components/schemas/PageVODepartmentApplicationSummaryVO"
        msg:
          type: string
        timestamp:
          type: integer
          format: int64
    ApplicationAnswerOptionVO:
      type: object
      properties:
        optionId:
          type: integer
          format: int64
        content:
          type: string
    ApplicationAnswerVO:
      type: object
      properties:
        questionId:
          type: integer
          format: int64
        questionTitle:
          type: string
        questionType:
          type: string
          enum:
          - SINGLE_CHOICE
          - MULTIPLE_CHOICE
          - SHORT_TEXT
          - LONG_TEXT
        answerText:
          type: string
        selectedOptions:
          type: array
          items:
            $ref: "#/components/schemas/ApplicationAnswerOptionVO"
    DepartmentApplicationDetailVO:
      type: object
      properties:
        campus:
          description: 学生所在校区；历史资料可为 null。更新资料时缺省或 null 表示不修改。
          type: [string, "null"]
          enum: [SOFTWARE_PARK, CENTRAL, QIANFOSHAN, XINGLONGSHAN, HONGJIALOU, BAOTUQUAN, null]
        id:
          type: integer
          format: int64
        departmentId:
          type: integer
          format: int64
        casId:
          type: string
        applicantName:
          type: string
        college:
          type: string
        major:
          type: string
        grade:
          type: integer
          format: int32
        phone:
          type: string
        email:
          type: string
        qq:
          type: string
        status:
          type: string
          enum:
          - SUBMITTED
          - INTERVIEW_COMPLETED
          - ADMISSION_DRAFT
          - ADMITTED
        submittedAt:
          type: string
          format: date-time
        answers:
          type: array
          items:
            $ref: "#/components/schemas/ApplicationAnswerVO"
    ResultDepartmentApplicationDetailVO:
      type: object
      properties:
        code:
          type: integer
          format: int32
        data:
          $ref: "#/components/schemas/DepartmentApplicationDetailVO"
        msg:
          type: string
        timestamp:
          type: integer
          format: int64
...
    ResultVoid:
      type: object
      properties:
        code:
          type: integer
          format: int32
        data: {}
        msg:
          type: string
        timestamp:
          type: integer
          format: int64
  securitySchemes:
    sessionCookie:
      type: apiKey
      name: JOINSESSION
      in: cookie